Make assoc_mgr running in slurmdbd fill in tres arrays for assocs with tres strings
For assocs with tres strings their related *_ctld tres arrays were only
filled in when assoc_mgr was running in slurmctld. This patch removes checks
in the following functions that limited them to function in slurmctld
only.

        assoc_mgr_set_tres_cnt_array()
        assoc_mgr_set_assoc_tres_cnt()

With this removal assoc_mgr running in slurmdbd will now fill in *_ctld tres
arrays for assocs with tres strings.

Ticket 18647