Allow recovery of a persistent connection when SLURM_PERSIST_INIT fails
On failure, the slurm_persist_rc_msg() value of "ret_info" can be the
value of the RPC that failed, not the protocol version.  In the case of
an error from SLURM_PERSIST_INIT, this is the case.  Since the unpack of
the rc messgae was successful, the protocol version would be overwritten
with the RPC type.

When opening a new connection, only set the protocol version if the
return code indicates a success to prevent this.

Ticket 19804