Improve MaxMemPerCPU check on exclusive jobs requesting per node memory.
After commit 215552c964, we use number of CPUs on the first node in the
partition as an estimate of maximum number of CPUs per node in the
partition. We already have a precise number that can be used here updated in
build_part_bitmap, which is part_ptr->max_cpu_cnt.

Ticket 19722