Don't calculate num_tasks for external launcher steps.
If --external-launcher and -c are used without -n leave num_tasks as NO_VAL
in _pick_step_nodes. It will set num_tasks to the number of selected nodes
after. This is to avoid hitting an xassert in _set_max_num_tasks() since
cpus_per_task is 0 for --external-launcher steps.

Ticket 18550