Allocate MPI port ranges to jobs if STEPMGR_ENABLED
Allocate a range of MPI ports to a job if a slurmstepd will be running
the step manager. Allocate the requested port count or allocate a port
count equal to the max task count on any node plus one.

In the case where the needed ports are busy on the selected nodes set the
job as pending.

Deallocate the port range when the job finishes.

Ticket 19652