Avoid underflow from old job flush event time
If the controller does a cold start with an older time stamped
node_state file (due to a slurmctld crash), job flush event time would
be older than the suspended job's time and cause an underflow and sql
error. To avoid the underflow always use the older suspend time.

Ticket 19841

Co-authored-by: Danny Auble <da@schedmd.com>