Reject salloc/sbatch using --resv-ports without slurmstepd step manager
The option --resv-ports can only be used with salloc and sbatch if the
slurmstepd step manager is enabled. It can be enabled either with --stepmgr
or SlurmctldParameters in slurm.conf. When a job is submitted without
the step manager enabled emit the following error:
"Slurmstepd step management must be enabled to use --resv-ports for job
allocations"

Ticket 19652