Save the cred state during reconfig of slurmd.
Because reconfig now starts up a new slurmd that reads in the save state,
running jobs that were added to cred_job_list could be lost since the cred
state was not always being save before reconfig. This issue could easily be
reproduce with PrologFlags=Alloc if reconfig happed between job allocation
and running the first step. This is because _rpc_prolog() adds the job to
cred_job_list without saving the cred state.

Ticket 19854