sacct - Fix printing of job group for job steps.
Previously, sacct was incorrectly displaying 4294967294 for the Group field
for job step entries. NULL (blank) should be displayed instead as it is done
with User field.

The slurmdb_step_rec_t struct has no uid or gid members because this is job
(not step) information, so it makes no sense to attempt to generate and
display the group string off of a non-existent job step gid member.

Ticket 19906