Improve default job resv_port_cnt calculations for stepmgr enabled jobs
This modifies how the defualt resv_port_cnt is calculated for jobs that are
using the stepd step manager. If a task count is not given and overcommit
is not used set the resv_port_cnt equal to the max cpus allocaed to any
given node plus 1. If overcommit is used the job_resrcs's tasks_per_node[]
values will all be 1 so estimate the max tasks on any given node plus 1.

Ticket 19652