scrun - Do not silently fail when parsing bad environment variables
Provide a useful log message instead of failing without any clue as to
why when an invalid environment variable causes the job submission to be
rejected. This was a problem for users which didn't know the cause
of scrun exiting, and just saw an error in the launcher (e.g. in docker).

Ticket 19943.