Fix assertion on restore of restoration of slingshot plugin
Prior to 166718092c the vni_table was created after restoring if there
was no restore. If a restore happens we unpack into the vni_table and
expect it to NULL but now with 28a57da052 the vni_table is created on
init -- so we assert on the vni_table not being NULL. So free it before
unpacking.

Ticket 19123