acct_gather_energy/ipmi,gpu,xcc - move _get_joules_task error to debug
When several threads send a request to slurmd to get the energy, we can
reach req_cnt limit of concurrent energy RPCs (10). When this happens the
error is logged to the user for each step where this happened. Fix this
extra verbosity by returning a new type of error (ESLURMD_TOO_MANY_RPCS),
so the caller (_get_joules_task), can detect this condition from
slurm_get_node_energy and log it as a debug message instead of an error,
thus reducing the annoying verbosity to the user.

Ticket 19779.