acct_gather_energy/gpu - Prevent a segfault on startup
Prevent a segfault on startup when running without GresTypes defined: the
gpu plugin is not initialized if GresTypes is not defined, and thus a call
to gpu_g_get_device_count() will crash in a non-developer build or abort in
a developer build.

If gres_get_gres_cnt() returns more than zero, then we know that the gpu
plugin is also initialized, so we can use gres_get_gres_cnt() to prevent
the crash. Because we are now using gres_get_gres_cnt() in the slurmd, not
just the slurmctld, move the function declaration in the header file to the
section with functions used by any daemon.

Ticket 19369