data_parser/v0.0.39 - Warn if parsing crontab field for job_desc_msg
This emits a warning if parsing crontab for PARSER_ARRAY(JOB_DESC_MSG).
It ignores the crontab entry. This is because the slurmrestd does not have
an endpoint that calls slurm_update_crontab() which is needed to submit a
cron job. So parsing this value is useless. This effects the following
endpoints:
POST /slurm/v0.0.39/job/{job_id}
POST /slurm/v0.0.39/job/submit

Ticket 18444