accounting_storage/mysql - use assoc rec valid_qos field to filter recs by qos
Getting/modifying/removing assocs using sacctmgr when a qos list was
supplied caused a 3-way sql join to happen on the assoc table to obtain
an assoc's inherited qos values. For a site with a sizable number of
assoc table entries, this would cause sacctmgr to hang while the database
server built an extremely large temporary table and then proceeded to
process it, consuming a large amount of cpu time.

Since the assoc_mgr already keeps a bit string of the inherited qos values
for an assoc, this patch leverages that by filtering assoc records within
slurmdbd using it rather than constructing an expensive sql query. This
change significantly speeds up the filtering and avoids the client hang
and load on the database server.

Ticket 18865