Join a new mount namespace in clone and mount /proc for get-user-env
Getting the user environment e.g. by setting --get-user-env or export=
in a batch script spawns a child which runs containerized into a pid
namespace, and which later executes a su or sudo to log in as the user
and get his environment. Some user environment scripts may require
reading /proc, and since these programs are run into its own pid
namespace starting from pid 1, they will also require its specific
/proc mountpoint or they would access the /proc related data from the host
and not from the container.

With this fix the cloned child is started in a new mount namespace.
In essence this emulates what "unshare --mount-proc" does, and fixes
issues where /proc/<pid> incorrectly resolved to processes from the host
namespace, which could make external init scripts to not behave as
they should (like when using lmod).

If the fork() approach is used instead of clone() (FreeBSD only), then we
don't need to mount /proc since we will not be working with namespaces.

Ticket 19160.