topology/block - Fix segfault in slurmctld
When doing the qsort to optimize the choose of a llblock by choosing
the biggest llblock inside a block, we were assuming the number of
nodes in llblock were all the same, so we were using an incorrect index
and getting out of boundaries.

Also you can configure more nodes in your block than the ones you really
have, and they will not be in the nodes_on_llblock[] bitmap. This was a
requirement of NVIDIA which might be useful for maintenance or down nodes.

Ticket 20103.