Fix mixed up IP addr forwarding with auth/slurm
When network credentials were added to auth/slurm in d6989b1, a dict was
used to pass the IP addresses in the net credentials. The unordered
nature of JSON dicts resulted in shuffled IP addresses and crossed wires
in fanout communication between slurmds. Add a new 'netcred' key to the
JWT token with an ordered list of node addrs. Keep the existing 'addrs'
key for backward compatibility.

Ticket 19366