Run burst_buffer.lua using script_launcher
Remove code that was previously used to run burst_buffer.lua; it duplicated
a lot of the functionality of run_command(), and was not async-signal-safe.
Replace it with run_command() using script_launcher, which is
async-signal-safe.

Pass the minimal slurm configuration needed to initialize the
burst_buffer/lua plugin. This includes information about logging so
burst_buffer.lua can log to the slurmctld log file at the correct log level
and with the correct log format. This also requires initializing the auth
and hash plugins: burst_buffer.lua can make RPCs to the slurmctld which
might call Slurm APIs, so it needs minimal information required to do that.

Also pass minimal information needed to run burst_buffer.lua. Pass the
argv of burst_buffer.lua on the argv of script launcher. Pass additional
information (function, job id, job info) over the pipe.

Use the existing _write_msg() function to write on the pipe to
script_launcher. Because we are not using eio in script_launcher, we need
new code to read the data from the pipe. A slurmscriptd RPC is used to
simplify writing/reading and packing/unpacking.

Use an environment variable SLURM_SCRIPT_CONTEXT to tell the slurmctld to
run in burst_buffer.lua mode.

Ticket 16126