run_command - Close the write pipe to the child sooner
Close the write pipe to the child immediately after it is used, before
calling run_command_poll_child(). This means that the pipe will be closed
before waiting for the child to finish. If an error happened during the
write, when the child tries to read the required data from the pipe, the
pipe will be closed and the child can exit.

Ticket 16126