Fix a continue statement that always led to breaking out of the loop
A "continue" statement within a "do ... while (...)" loop only continues
execution of the loop body if the loop continuation condition is still true.
Since the condition will never be true in a "do ... while (false)" loop, the
"continue"statement has the same effect as a "break" statement.

Change the loop to be a while(true) and break at the end so that continue
always starts execution at the top of the loop.

CID 371043