acct_gather_energy/ipmi - Fix IPMI sensor info gathering
When EnergyIPMIPowerSensors is not defined in acct_gather.conf, readings
from the first power sensor returned by ipmi-sensors should be the default.
This was not the case and no sensors were used. This fix updates the number
of sensors detected when calling the energy gathering so that even when none
are defined per configuration, it will use the first sensor found.

Ticket 19415.