Fix segfault in stepgmr jobs when submitting within existing allocation
get_job_env() returns one allocated string and the array points to
different parts in the array. So when env_array_overwrite() was
overwriting an existing SLURM_STEPMGR env variable, it was asserting
because the pointer wasn't actually an xstring. _split_env() was created
to make the array an array of individual xstrings so it env variables
could be replaced. So _split_env() just needs to be called before adding
the SLURM_STEPMGR env variable.

Ticket 20135