foreach_job_by_id - Improve support for array expressions
The previous code was intended to call null_callback once for every
individual task that was not found. However, it only did this for the first
task, then returned SLURM_ERROR on FOR_EACH_JOB_BY_ID_EACH_CONT, which
resulted in breaking from the list_for_each_loop.

Rather than fix this, this commit improves the design:
Only call the null_callback once with all array tasks that were not found
set in array_bitmap, rather than once for every individual array task that
was not found.

Ticket 19776

Signed-off-by: Nathan Rini <nate@schedmd.com>