Adapt old MpiParams logic that assumed only a possible value
MpiParams until now could only contain the "ports=#-#" parameter.
Before adding a new parameter, we have to ensure that the old logic
does not assume that it can only contain this.

This change mantains the old general logic, but instead of checking
that there is something in MpiParams, it checks that there is a
"ports=..." parameter present.

Ticket 20115.