scancel --ctld - Fix canceling the final pending array task
For scancel --ctld and the associated rest api endpoints, when trying to
cancel only the meta job and all other tasks have been split out, the meta
job was not cancelled. This is an issue when filtering jobs in
_apply_signal_jobs_filter(). Before this commit, we put the meta record in
the pending_array_task_list. we called _foreach_signal_job_array_tasks()
on that list. That function calls _signal_pending_job_array_tasks() which
removes tasks from the meta record but does not actually cancel the meta
record. We need to call job_signal() on the meta record to cancel it.

Fix this by checking the meta record's array_recs->task_cnt. If that is
zero, then all tasks have been split from the meta record, meaning we should
call job_signal() on the meta record to cancel that job.

Ticket 20195