Fix setting limits to 0 when when modifying the flags on a Slurm account
When altering the flags on a Slurm account (UsersAreCoords) several limits
on the account's association would be set to 0 in Slurm's internal cache.

The values in the database would remain unchanged, only the values used
by the slurmctld would be set to 0 on the association (MaxJobs being one
of them) and would result in jobs being denied. Without this patch
restarting the slurmctld would fix the erroneous cache.

Ticket 20214