Set errno in slurm_persist_conn_open() if no response to persist_init()
Previously, in the case that there was no response to persist_init() it
would leave errno as 0. This would lead to error 'No error' logs:
"error: Sending PersistInit msg: No error"
Additionaly, the persistent connection would countinue to be used despite
the fd being closed. For example, slurm_rest_auth_p_get_db_conn() used the
value of errno to determine if the db_conn should be set to NULL or not.

Ticket 18516