Allow persistent connection rc message to be sent back on error.
slurm_persist_conn_process_msg() creates an error message if it fails to
unpack presist_msg. This was never able to be sent to the caller though
on if it was a presist_init message because presist_conn->tls_conn would be
NULL. It only got initialized if the unpack was successful.

Now if there is an error unpacking the presis_msg it will create a tls_conn
if it doesn't have one.

Ticket 18516