Clarify slurmrestd's failed to open slurmdbd connection error message
Now if errno is set to SLURM_PROTOCOL_AUTHENTICATION_ERROR add the
following clarification to the error: ", authentication error".
For example, "Failed to open slurmdbd connection, authentication error", or
"Failed to open connection to slurmdbd, authentication error. Response
fields may not be fully populated or empty."
This change is intended to make it more obvious to the user that they are
dealing with an authentication issue, like an expired token, instead of
implying that the slurmdbd is not reachable.

Ticket 18516