Fix controller regression not deleting job hash dirs after MinJobAge
A regression was caused by bb987a7f61 inadvertently removed the
code that put the jobid on the purge_files_list which would delete the
job hash dirs. The problem is that purge_files_list is only available to
the controller. So make it local to job_record_.h and only use it when
the controller is deleting a job_record_t. The controller through
_copy_job_desc_to_file() is the only one writing the job script and
environment to the file system.

Ticket 20238