acct_gather_energy/ipmi,gpu,xcc - change debug to log_flag
Continuation of commit b940bc9750.

The message for the case where req_cnt > 10 for the _rpc_acct_gather_energy
in the step side (_get_joules_task function) was changed to debug. However,
it is more appropiate to output this message through the Energy debug flag.

NEWS have been adapted.

Ticket 19779.