Add /slurm/{version}/nodes POST endpoint to enable multinode update
Previously only a single node was supported, encoded as a URL param:

curl \
  -H 'Content-Type: Application/json' \
  -H X-SLURM-USER-TOKEN:$SLURM_JWT \
  http://slurmrestd.site/slurm/v0.0.42/node/n1> \
  --data-binary '{"state": ["drain"], "reason": "failing disk"}'

As requested, the following is now possible:

curl \
  -H 'Content-Type: Application/json' \
  -H X-SLURM-USER-TOKEN:$SLURM_JWT \
  http://slurmrestd.site/slurm/v0.0.42/nodes \
  --data-binary '{"name": ["n[1-6]"], "state": ["resume"]}'

Note that for this body parser, "name" is expected to be an array and
so the square brackets are required around the nodename expression.

However, the parser *will* accept a plain string and work as expected,
logging only a warning. The following are thus possible and correct:

curl \
  [...]
  --data-binary '{"name": ["n[1-6],n12"], "state": ["resume"]}'

curl \
  [...]
  --data-binary '{"name": "n[1-6]", "state": ["resume"]}'

Ticket 19745