slurmd - Fix request counter underflow for energy gathering
This is a regression introduced by commit c9789fe0c4e.

When the number of threads requesting energy exceeds 10 it is possible to
reach a condition where some threads do not increment req_cnt but all of
them decrement the variable. This ends up in an underflow of this variable,
making it impossible to gather energy for any thread until a slurmd reboot.

Ticket 19779.