salloc - Always display node names when allocation is granted
The "Nodes <nodelist> are ready for job" message when performing a salloc
was only displayed if salloc actually had to wait for the nodes to be ready
for more than one iteration. This message should appear regardless of that,
so it will now always be printed if the allocation is granted and there are
no other issues.

Ticket 20209.