Sync job_resrcs->cpus_used with the number of used cores
Modify the number of cpus added or removed from job_resrcs_ptr->cpus_used[]
to be a multiple of threads per core, except in some cases where CR_CPU is
used and threads per core is > 1, which is not recommended. If CR_CPU is
used and the job is not allocated a multiple of thread per core cpus don't
set cpus_used[] more then the allocated cpus[] on that node. Additionally,
when deallocating remove thread per core units from cpus_used instead of
partial cores.

This change keeps the number of used_cpus in sync with the allocated core
bitmap. Before only the number of requested cpus would be tracked as used
despite the fact full cores are always allocated to steps.

This does not change the accounting of cpus for jobs or steps. Which for
CR_CPU currently only shows the number of cpus requested and not actual
allocated which is cores * threads per core. This is also true for CR_CORE
but only for steps.

Ticket 19960