Fix QOS preemption with PreemptMode=SUSPEND
When we run _job_test in RUN_NOW mode to start a job that suspends other
jobs based on QoS we were just using an empty row for CPU selection.
This mechanism ignored jobs already running on the cluster - it was just
selecting resources based on _select_nodes rules applied to idle cluster.

This commit adds a loop checking if the job WILL_RUN if it suspends
resources from preemptee candidates, without actually suspending those jobs.
If the job can run by preempting those jobs, pass the list of the jobs that
need to be suspended to _job_test and then select resources in addition to
cores from those jobs removed.

The code doesn't allow start of job that will mix CANCEL/REQUEUE mode of
preemption with SUSPEND.

The decision to attempt SUSPEND vs CANCEL/REQUEUE is arbitrary as it was
before the commit. It comes from the suspension mode of the first job on
preemptee_candidates list.

Ticket: 20023