conmgr - Fix race resulting in the process not shutting down
Do not cache mgr.shutdown in a local variable. mgr.mutex is unlocked and
then locked again, so the cached value may no longer be valid. If
mgr.shutdown_requested was false when the worker woke up, then the worker
did not signal watch to wake up. This could lead to watch never waking up,
resulting in the process never shutting down.

Ticket 20114

Signed-off-by: Nathan Rini <nate@schedmd.com>