conmgr - Do not revert signal handlers in developer mode
Reverting signals in developer mode meant that behavior was different in
developer and non-developer mode. If there was no signal handler before
installing the signal handler with conmgr, the default signal handler was
used. For example, assume conmgr added signal work for SIGTERM that called
conmgr_request_shutdown(). The first time the process received SIGTERM, the
conmgr_shutdown() was run. This would result in conmgr reverting the signal
handlers (if in developer mode). So the second time the process received
SIGTERM, the process would terminate because that is the default behavior
for SIGTERM.

Ticket 20114

Signed-off-by: Nathan Rini <nate@schedmd.com>