conmgr - Always write to unpollable connections
Assume unpollable connections can always be written. The kernel will
return EWOULDBLOCK or EAGAIN if the connection is not ready for a write
which will cause handle_write() to yield and watch to retry the write.
This is not ideal but in theory this only applies to files and the kernel
will manage the write buffer outside of the view of conmgr.

Ticket 20114, 20426