Add missing bitflag information to job_desc for _valid_pn_min_memory
The job_desc_msg_t parameter for the _valid_pn_min_memory function
was missing the job's bitflags. The bitflags are necessary to determine
if the job was requested with a specific cpus-per-task or not, using
JOB_CPUS_SET. This is a preliminary change needed for the next commit.

Ticket 19335