Respect CPUs/Task when requested and memory constraints are not met
When using --mem-per-cpu in allocation commands, if the requested value
exceeds the configured DefMemPerCPU or MaxMemPerCPU in slurm.conf, it
would automatically increase the amount of allocated CPUs to meet those
memory requirements.

This was done by internally changing the job's cpus-per-task value,
which could be unexpected by the user if a specific value was already
requested with the -c option. This change addresses this by directly
increasing the amount of requested CPUs (without modifying cpus-per-task)
if the user has already specified a cpus-per-task value.

Ticket 19335