Fix GPU env vars w/ allow-task-sharing to be within job context
The --gres-flags=allow-task-sharing option lets all tasks on a node see all
GPUs within the job on the node. Previously, the indices for these env vars
were set within the context of the node, not the context of the job. This
meant that the IDs set for the GPUs in the environment variables could
seemingly be out of range.

This commit consolidates the bits in gres_bit_alloc (gres allocated to job)
in case they aren't already consolidated (e.g. all GPUs on node are already
allocated to the job) which will effectively result in the correct indices
for environment variables.

Ticket 20433