conmgr - Update comment for conmgr_run()
Note that conmgr_run() will run until it is stopped via explicit request.
Make this change in behavior for conmgr from 24.05 explicit. Previously,
conmgr would run until it had no more connections or work to do. This was
convenient for inetd mode in slurmrestd but it is also the only use case
and can easily be modified to ask conmgr to shutdown instead. Keeping
conmgr running until shutdown makes it easier to reason about instead of
trying to handle all of the race conditions with signal mgr during
shutdown.

Ticket 20114, 20426