Add "curr_dependency" and "orig_dependency" to the Lua job record
Although the job_details_t struct contains "dependency" and
"orig_dependency", we do not add "dependency" to the Lua job record to avoid
potential confusion.

An admin might want to use "dependency" to get the original dependency of
the job. However, in the job_details_t struct, "dependency" is used for the
current dependency and "orig_dependency" is used for the original
dependency.

The original dependency is in the jobcomp/elasticsearch and kafka plugins.
See the function jobcomp_common_write_state_file() in
src/plugins/jobcomp/common/jobcomp_common.c.

The original dependency is also in the jobcomp/script plugin. See
_jobcomp_info_create() in src/plugins/jocomp/script/jobcomp_script.c.

To prevent confusion by an admin writing a jobcomp.lua script, avoid the use
of the possibly ambiguous "dependency" field for the job record. Instead,
use "curr_dependency" and "orig_dependency" to represent the current and
original dependency values, respectively.

Adding these fields to the Lua job record table means that the
job_submit.lua plugin will also be able to use these values in the
slurm_job_modify() hook.

Ticket 20292