conmgr - Switch to quiesced dependency work
New model to have work that has dependency to run when conmgr is fully
quiesced as opposed to requesting quiesced state with a blocking wait for
it with conmgr_quiesce(). watch() can then run the quiesced work when
ready avoiding multiple race conditions of waiting for workers to stop.

conmgr_quiesce() changed to stub to avoid breaking compiles.

Ticket 20114