Add new parameter to child_args to determine when to use mount calls
Since there now will be situations where we might not want to perform
mount calls in the child process responsible for retrieving the user
environment, we need to be able to indicate this from the parent process.
This commit adds a new parameter to the child_args struct to control this.

Ticket 19286