Add new _ns_disabled() function
This function checks that there are no PID and mount namespace limits
in place. In this context, having namespace limits implies that the
contents of /proc/sys/user/max_[pid|mnt]_namespaces are 0, effectively
disabling them.

This is added as an auxiliary function to ensure that namespace creation
is available before trying to execute clone and mount calls that rely
on it.

Ticket 19286