Fall back to fork() if there are namespace limits when retrieving env
The change to using clone() + mount() calls instead of fork() to spawn
the child process responsible for retrieving user environments
(5c3810652c) made namespace creation mandatory. If the user had this
mechanism disabled, there was no fallback to the old behavior.

This commit adds logic to check if namespaces are available, and if not,
reverts back to the old fork-based logic. The implications of this
fallback are specified as a warning through the debug level logs.

Ticket 19286