slurmd - _increment_thd_count before accept
Check if we can create new thread before accepting connection.
Otherwise, if we already completed _wait_for_all_threads and are about
to reconfigure/restart the accepted message will be lost. If we leave in
the fd new slurmd may pick it and server after reconfigure.

Since we increase the number of active threads before accept we cannot
hang on it. Instead wait on poll and call accept only if we know that
there is data to read.

Ticket 19563