Do int type cast in case rem_cpus has underflowed
*switch_cpu_cnt is uint32_t, so it needs to be cast to int before being
compared to rem_cpus in case rem_cpus has underflowed, which is a valid
state. e.g. when we are getting exclusive node allocations, the job will
reach the CPU requirement (rem_cpus) sooner than it reaches the node
count requirement (rem_nodes).

Note that this is not an issue for switch_node_cnt since it is already
type int.

Ticket 20547