Ignore --json/yaml for scontrol show config to prevent mixing output types.
Before scontrol show config --json would print the config as if --json was
not used, but it would print the ping as json.

There isn't a parser for slurm.conf, but there is one for ping, which gets
printed as part of _print_config(). So to make sure the ouput is not mixing
output types, disable json/yaml ouput for ping by setting mime_type NULL.

Ticket 18156