Add new option bf_allow_magnetic_slot to SchedulerParamters
By default, jobs in a magnetic reservation are not added in the
backfill plan. Which can lead to some jobs not respecting priorities
inside magnetic reservations.

This new option will allow magnetic jobs to get a slot in the
backfill plan, preventing lower priority jobs skipping priority
order in that situation.

Ticket 19507