Fix allowing access to reservations without MaxStartDelay set
Currently if --signal=R:<sig_num>@0 is used all reservations that do not
have MaxStartDelay set will have their nodes available to the job. This
is because MaxStartDelay's default value is none which is 0. The check to
see if the nodes of a reservation can be used only checks if the warn_time
of the job is less than MaxStartDelay and if the 'R' option was used. In
the case that warn_time (@sig_time) is 0 and the MaxStartDelay is not set,
i.e. 0, the job will consider the reservation's nodes.

This is prevented by making sure MaxStartDelay is not none, 0, first before
allowing the nodes to be considered.

Ticket 20564