Fix reboot asap nodes being considered in backfill after a restart
"scontrol reboot asap <node>" defaulted the node's next_state to
"resume" without putting it in the rs_node_bitmap. On a restart,
_build_bitmaps() would put any node with next_state=resume into the
rs_node_bitmap when it should only happen if explicitly requested with
next_state=resume.

This also solves the problem of not being able to distinguish between
explicitly setting next_state=resume and an asap reboot. Now asap
reboots will show a next_state=UNDRAIN if asap is not combined with an
explicit next_state request.

Ticket 20596

Signed-off-by: Joel Criado <joel@schedmd.com>