Re-exec slurmctld to run slurmscriptd
conmgr_init() creates threads and is called before slurmscriptd_init().
slurmscriptd_init() calls fork() and the slurmscriptd runs in the child
process. fork(2) says:

       After a fork() in a multithreaded program, the child can safely call
       only async-signal-safe functions (see signal-safety(7)) until such
       time as it calls execve(2).

In order to comply with this, re-exec slurmctld after the fork() in
slurmscriptd mode which is denoted by an environment variable. The newly
exec'd slurmctld re-initializes its configuration and then runs the
slurmscriptd main loop.

Ticket 20114

Signed-off-by: Nathan Rini <nate@schedmd.com>