Stop backfill from planning jobs that overlap in the same time slot
Before, if a job's time limit was less than bf_resolution, it was possible
for it to be scheduled in the same time slot reserved for another job using
the same nodes. This happened if the start_time was equal to end_reserve,
which was possible because both were rounded down to the nearest
bf_resolution.

This makes sure that end_reserve is always at least bf_resolution more than
start_time.

Ticket 20564