conmgr - Rework how conmgr configuration parameters are applied
Previously it was expected that conmgr_init() would be called before
applying configuration parameters. This was problematic because
conmgr_init() did not allow for decreasing the number of workers or
connections.

This commit changes the paradigm: Set the configured values first, then use
the configured values when initializing conmgr.

Ticket 20114