Fix user look up using possible unrealized uid in the dbd.
The dbd calls assoc_mgr_set_missing_uids() each hour hoping
to fill in any missing uids on users added without having been added
to the system yet. This causes up to an hour wait unless you restart
the daemon.

Because of this we should only ever use the user name to look the user
up in the assoc_mgr. We already do this elsewhere (though it is not
spelled out) _foreach_add_user() in as_mysql_user.c.

Ticket 20142

Signed-off-by: Connor Moore <conner@schedmd.com>