Fix replacement/creation of reservations when overlapping other reservations
Reservations should be allowed to replace nodes that are being part of
other reservations with the maint or overlap flags set. The same
happens for reservation creation.

There was an incorrect flag probably coming from a bad copy-paste that
made the two bitmaps SELECT_OVR_RSVD and SELECT_NOT_RSVD to be equal.

Ticket 20803.