Docs - Clarify -c without -n when --mem-per-cpu exceeds MaxMemPerCPU
If -c and --mem-per-cpu is used, but --mem-per-cpu exceeds MaxMemPerCPU,
the number of allocated cpus will be increased. If -n is not used the
max number of task that can be executed will be increased based off of the
cpus allocated. This leads to more tasks then what the user expected.

Ticket 20583