Read the MaxRSS memory value from tres_data[].size_write
Only for TRES_ARRAY_MEM we will store the new memory peak metric received
from the accounting plugins into a tres_usage_in_max[].

So we first process tres_usage_in_tot to not lose size_read, and then we
overwrite size_read with the size_write value. After that the code will
take care of tres_usage_in_max.

If the new metric did not exist then we will fallback to previous behavior
and we will store the instant readings instead of the memory.peak.