Avoid underflow errors when calculating the number of gres for steps
If less gres were requested than nodes in an allocation, the srun would
fail, throwing the error saying there was insufficient gres. This is
because of two underflows in max_gres and in gres_needed. The max_gres
caused too many gres to be allocated which then caused the underflow in
gres_needed. If there are more gres requested than nodes, nodes will be
packed instead, and will not request more gres than required.

Ticket 20591
--
2.43.0