Fix cgroup/v2 cleanup when running inside a container
We were leaving stray cgroups behind when we could not do the cleanup.
This fixes this situation by ensuring the slurmstepd pid can be moved out to
a different cgroup than the one we're deleting before trying to delete it.

We put the process into the system cgroup instead of to the root cgroup.
The system cgroup is also a leaf so we are not violating the
no-internal-constrain kernel rule.

In cgroup/v1 this is already done as we move the process to the subsystem
directory (where we will always have permissions) instead of the root.

Specially needed for containerized environments where /sys/fs/cgroup might
not be writeable by us.

Ticket 20764.