Make SchedSubmit always consider priority tiers in _top_priority()
This fixes a regression added in commit ab663ba which made it so if both
jobs did not have a reservation only the priority of the jobs would be
considered. Now if both do not have a reservation burst buffer stage-in and
partition priority tiers are also considered.

Additionally, this makes it so that  burst buffer stage-in and partition
priority tiers are always considered when reservations match. Before only
the job priority was being considered.

Lastly, before only the first reservation specified by the user would be
considered. However, users can specify multiple reservations. If any
reservation between the two jobs overlap consider the jobs as using the
same reservation.

Ticket 20893