cgroup/v2 - Check for cgroup interface access
There are some kernels, (e.g. 4.x series in RHEL8) that do not support
certain cgroup interfaces. One of them is memory.peak in cgroup/v2.

In order to not emit unnecessary errors and avoid constant checks for
these interfaces, we detect the existence and access permissions to these
interfaces at the beggining. A similar solution could be implemented in
cgroup/v1 if this were the case, but at this time all the interfaces we
use are supported since forever.