Prevent the main scheduler from skipping jobs using FLEX reservations
Before the main scheduler could not schedule a job if it was requesting a
FLEX reservation and the reserved nodes overlapping with the partition was
not enough for the job. This is a problem because jobs in FLEX reservations
have access to all nodes, not just the reserved nodes.

Now return true from _job_runnable_test3() if the job is using a FLEX
reservation, which will allow the main scheduler to see if the job can
run.

Ticket 20936