Partially fail by part for FLEX and ANY_NODES resv in main scheduler
Before, if a job using a reservation failed by part, i.e.
fail_by_part = true, it would always set fail_by_part = false and mark the
reservation as not being usable for further jobs. This presents a problem
for FLEX and ANY_NODES reservations since those reservations allow their
jobs to use nodes that are not specifically reserved. Not failing by part
allows jobs without reservations to be started on nodes that are needed by
the job in the reservation, thus delaying them.

This commit makes it so fail_by_part stays true if the reservation is FLEX
or ANY_NODES. It also sets fail_by_part_non_reserve to true so that only
nodes not in a reservation in the partition will be removed from the
avail_node_bitmap. The reserved nodes will be removed once the first job
not using a reservation gets evaluated in that partition. That first
non-reservation job cannot be scheduled since all the non-reserved nodes
have been removed and because the partition has been marked as
PART_FLAG_SCHED_FAILED.

Ticket 20936